Implement desync and clean-up: v0.1.10
This update simplifies a lot of the dialog handling even further, and adds the desynchronization functionality. The NPCs were simplified as well, and several classes were renamed to better fit their purpose.

+++ Orikivo/src/Desync/Mapping/Generation/ITableEntry.cs
@@ -2,6 +2,6 @@
 {
     public interface ITableEntry
     {
-        float Weight { get; set; }
+        int Weight { get; set; }
     }
 }
